requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0

# Data collection (local)
snscrape>=0.4.0

# Machine learning and evaluation
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Text processing and sentiment analysis (for baseline)
textblob>=0.17.1
vaderSentiment>=3.3.2

# Note: For Google Colab, install these separately:
# pip install "unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git"
# transformers>=4.35.0
# torch>=2.0.0
# accelerate>=0.24.0
# bitsandbytes>=0.41.0

